Extract useUnreadReactivity hook (simplify pass)
Both GroupPage and GroupList had a verbatim copy of the badge-settings +
poll-views subscription block. Lift it into lib/unread.ts so the event set
can't drift between the two surfaces.
